refactor(checkBalance/spendTokens): move to models dir
