+++ api/app/clients/OpenAIClient.js
@@ -3,7 +3,9 @@ const ChatGPTClient = require('./ChatGPTClient');
 const BaseClient = require('./BaseClient');
 const { getModelMaxTokens, genAzureChatCompletion } = require('../../utils');
 const { truncateText, formatMessage, CUT_OFF_PROMPT } = require('./prompts');
-const { createLLM, RunManager, checkBalance, spendTokens } = require('./llm');
+const checkBalance = require('../../models/checkBalance');
+const spendTokens = require('../../models/spendTokens');
+const { createLLM, RunManager } = require('./llm');
 const { summaryBuffer } = require('./memory');
 const { runTitleChain } = require('./chains');
 const { tokenSplit } = require('./document');
