+++ api/app/clients/llm/index.js
@@ -1,11 +1,7 @@
 const createLLM = require('./createLLM');
 const RunManager = require('./RunManager');
-const spendTokens = require('./spendTokens');
-const checkBalance = require('./checkBalance');
 
 module.exports = {
   createLLM,
   RunManager,
-  spendTokens,
-  checkBalance,
 };
